code: extract SALSSA algorithm implementation
