+++ lib/Transforms/IPO/CMakeLists.txt
@@ -6,7 +6,8 @@ add_llvm_pass_plugin(LLVMNextFM
   MergeSimilarFunctions.cpp
   MinSizeAttr.cpp
   MultipleSequenceAlignment.cpp
-  NextFMLoader.cpp)
+  NextFMLoader.cpp
+  SALSSACodeGen.cpp)
 
 if(NOT LLVM_ENABLE_RTTI)
   target_compile_options(LLVMNextFM PRIVATE "-fno-rtti")
